Moved services from Domain to API
Main reason is that these services are associated only with this API and
using the same DTOs as API controllers, so there's no need
to keep them in domain, because domain should be isolated from API logic
