+++ src/FoodDiary.API/Controllers/v1/ExportsController.cs
@@ -5,8 +5,8 @@ using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 using AutoMapper;
+using FoodDiary.API.Services;
 using FoodDiary.Domain.Dtos;
-using FoodDiary.Domain.Services;
 using FoodDiary.PdfGenerator;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
