+++ src/FoodDiary.API/Services/IProductService.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using FoodDiary.Domain.Dtos;
 using FoodDiary.Domain.Entities;
 
-namespace FoodDiary.Domain.Services
+namespace FoodDiary.API.Services
 {
     public interface IProductService
     {
