+++ src/FoodDiary.API/Services/Implementation/CategoryService.cs
@@ -6,9 +6,8 @@ using System.Threading.Tasks;
 using FoodDiary.Domain.Dtos;
 using FoodDiary.Domain.Entities;
 using FoodDiary.Domain.Repositories;
-using FoodDiary.Domain.Services;
 
-namespace FoodDiary.Infrastructure.Services
+namespace FoodDiary.API.Services.Implementation
 {
     public class CategoryService : ICategoryService
     {
