+++ tests/FoodDiary.UnitTests/Controllers/ImportsControllerTests.cs
@@ -4,8 +4,8 @@ using AutoFixture;
 using FluentAssertions;
 using FoodDiary.API.Controllers.v1;
 using FoodDiary.API.Options;
+using FoodDiary.API.Services;
 using FoodDiary.Domain.Dtos;
-using FoodDiary.Domain.Services;
 using FoodDiary.UnitTests.Customizations;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
