+++ tests/FoodDiary.UnitTests/Controllers/NotesControllerTests.cs
@@ -5,9 +5,9 @@ using AutoMapper;
 using FluentAssertions;
 using FoodDiary.API;
 using FoodDiary.API.Controllers.v1;
+using FoodDiary.API.Services;
 using FoodDiary.Domain.Dtos;
 using FoodDiary.Domain.Entities;
-using FoodDiary.Domain.Services;
 using FoodDiary.Domain.Utils;
 using FoodDiary.Infrastructure.Utils;
 using FoodDiary.UnitTests.Customizations;
