+++ tests/FoodDiary.UnitTests/Controllers/ProductsControllerTests.cs
@@ -6,9 +6,9 @@ using AutoMapper;
 using FluentAssertions;
 using FoodDiary.API;
 using FoodDiary.API.Controllers.v1;
+using FoodDiary.API.Services;
 using FoodDiary.Domain.Dtos;
 using FoodDiary.Domain.Entities;
-using FoodDiary.Domain.Services;
 using FoodDiary.UnitTests.Customizations;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
