+++ tests/FoodDiary.UnitTests/Services/CategoryServiceTests.cs
@@ -2,12 +2,12 @@
 using AutoFixture;
 using AutoFixture.Xunit2;
 using FluentAssertions;
+using FoodDiary.API.Services;
+using FoodDiary.API.Services.Implementation;
 using FoodDiary.Domain.Abstractions;
 using FoodDiary.Domain.Dtos;
 using FoodDiary.Domain.Entities;
 using FoodDiary.Domain.Repositories;
-using FoodDiary.Domain.Services;
-using FoodDiary.Infrastructure.Services;
 using FoodDiary.UnitTests.Customizations;
 using Moq;
 using Xunit;
