+++ tests/FoodDiary.UnitTests/Services/ExportServiceTests.cs
@@ -3,10 +3,10 @@ using System.Linq;
 using System.Threading;
 using AutoFixture;
 using FluentAssertions;
+using FoodDiary.API.Services;
+using FoodDiary.API.Services.Implementation;
 using FoodDiary.Domain.Entities;
 using FoodDiary.Domain.Repositories;
-using FoodDiary.Domain.Services;
-using FoodDiary.Infrastructure.Services;
 using FoodDiary.UnitTests.Customizations;
 using Moq;
 using Xunit;
